refactor: rename MrtDatasetTable to DatasetTable, remove old DataGrid-based DatasetTable
